Refactor mostly auditwheel
