+++ Cargo.toml
@@ -32,7 +32,6 @@ keyring = { version = "1.0.0", optional = true }
 platform-info = "0.2.0"
 pretty_env_logger = { version = "0.4.0", optional = true }
 regex = "1.4.5"
-rpassword = { version = "5.0.1", optional = true }
 serde = { version = "1.0.131", features = ["derive"] }
 serde_json = "1.0.70"
 sha2 = "0.10.0"
@@ -42,8 +41,7 @@ tempfile = "3.2.0"
 toml = "0.5.8"
 zip = "0.5.5"
 thiserror = "1.0.24"
-dirs = { version = "4.0.0", optional = true }
-configparser = { version = "3.0.0", optional = true }
+dirs = "4.0.0"
 fs-err = "2.5.0"
 fat-macho = { version = "0.4.4", default-features = false }
 once_cell = "1.7.2"
@@ -61,15 +59,18 @@ clap = { version = "3.0.0", features = ["derive", "env", "wrap_help"] }
 clap_complete = "3.0.0"
 clap_complete_fig = "3.0.0"
 semver = "1.0.4"
+# upload
+configparser = { version = "3.0.0", optional = true }
+multipart = { version = "0.18.0", features = ["client"], default-features = false, optional = true }
+rpassword = { version = "5.0.1", optional = true }
 ureq = { version = "2.3.1", optional = true }
-multipart = { version = "0.18.0", features = ["client"],default-features = false, optional = true }
 
 [dev-dependencies]
 indoc = "1.0.3"
 
 [features]
 default = ["log", "upload", "human-panic"]
-upload = ["ureq", "multipart", "rpassword", "configparser", "dirs"]
+upload = ["ureq", "multipart", "rpassword", "configparser"]
 password-storage = ["upload", "keyring"]
 log = ["pretty_env_logger"]
 # Internal feature to speed up the tests significantly
