+++ src/auditwheel/mod.rs
@@ -8,4 +8,4 @@ mod repair;
 pub use audit::*;
 pub use platform_tag::PlatformTag;
 pub use policy::{Policy, MANYLINUX_POLICIES, MUSLLINUX_POLICIES};
-pub use repair::{find_external_libs, hash_file};
+pub use repair::find_external_libs;
